pyproject.toml
[project]
name = "demand-predict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py==2.1.0",
    "aiohappyeyeballs==2.4.8",
    "aiohttp==3.11.13",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "attrs==25.1.0",
    "captum==0.7.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "docker-pycreds==0.4.0",
    "einshape==1.0",
    "filelock==3.17.0",
    "fonttools==4.56.0",
    "frozenlist==1.5.0",
    "fsspec==2025.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "grpcio==1.70.0",
    "holidays==0.68",
    "huggingface-hub==0.29.1",
    "idna==3.10",
    "inquirerpy==0.3.4",
    "jax==0.5.1",
    "jaxlib==0.5.1",
    "jinja2==3.1.5",
    "joblib==1.4.2",
    "kaleido==0.2.1",
    "kiwisolver==1.4.8",
    "lightning-utilities==0.12.0",
    "markdown==3.7",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.1",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.1",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "narwhals==1.29.0",
    "networkx==3.4.2",
    "neuralprophet==0.9.0",
    "numpy==2.2.3",
    "opt-einsum==3.4.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pfzy==0.3.4",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "plotly==6.0.0",
    "prompt-toolkit==3.0.50",
    "propcache==0.3.0",
    "protobuf==5.29.3",
    "psutil==7.0.0",
    "psycopg2-binary==2.9.10",
    "py4j==0.10.9.7",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "pyparsing==3.2.1",
    "pyspark==3.5.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytorch-lightning==2.5.0.post0",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rich==13.9.4",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentry-sdk==2.22.0",
    "setproctitle==1.3.5",
    "setuptools==75.8.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sqlalchemy==2.0.38",
    "sympy==1.13.1",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "threadpoolctl==3.5.0",
    "timesfm==1.2.8",
    "torch==2.6.0",
    "torchmetrics==1.6.2",
    "tqdm==4.67.1",
    "typer==0.15.2",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "utilsforecast==0.2.12",
    "wandb==0.19.7",
    "wcwidth==0.2.13",
    "werkzeug==3.1.3",
    "yarl==1.18.3",
]
